refactor: `Player`를 역할에 맞게 `MoneyLottosDto`로 수정
