+++ src/main/java/lotto/controller/dto/MoneyLottosDto.java
@@ -1,20 +1,20 @@
-package lotto.model.domain;
+package lotto.controller.dto;
 
 import lotto.model.domain.vo.Lottos;
 import lotto.model.domain.vo.Money;
 
-public class Player {
+public class MoneyLottosDto {
 
     private Money money;
     private Lottos lottos;
 
-    private Player(Money money, Lottos lottos) {
+    private MoneyLottosDto(Money money, Lottos lottos) {
         this.money = money;
         this.lottos = lottos;
     }
 
-    public static Player from(Money money, Lottos lottos) {
-        return new Player(money, lottos);
+    public static MoneyLottosDto from(Money money, Lottos lottos) {
+        return new MoneyLottosDto(money, lottos);
     }
 
     public Money getMoney() {
